Move the flask server for ui to new main file that will be executed
